Pixel formats refactor/update:  - Added 8/16/32 bit integer formats  - Added 8/16 bit signed normalized formats  - Added 16 bit unsigned normalized formats  - Renamed old pixel formats to follow a nicer naming convention  - Updated pixel packing/unpacking code so it is more generalized
Former-commit-id: a2cbf9edddc45ce8d3b6ce578a23bf63fb828b12

+++ Source/RenderBeast/Source/BsLightProbes.cpp
@@ -200,7 +200,7 @@ namespace bs { namespace ct
 				for(UINT32 i = 0; i < 4; ++i)
 				{
 					// Check for outer vertices, which have no SH data associated with them
-					if (entry.volume.vertices[i] >= innerVertexCount)
+					if (entry.volume.vertices[i] >= (INT32)innerVertexCount)
 						entry.volume.vertices[i] = -1;
 					else
 						entry.volume.vertices[i] = mTempTetrahedronBufferIndices[i];
